+++ examples/common.ts
@@ -10,8 +10,8 @@ import {
   result,
   sepBy,
   sequence,
-  updatePosition,
 } from "../src/index.ts";
+import { updatePosition } from "../src/utilities.ts";
 
 /**
  * Represents a predicate function
